Externalizing and documenting PropertiesBuilder and StepsBuilder class.
